+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.0.0
+
+* New transition animation for button width, background color, and icon color
+* To disable the animation, set `animationDuration` to `Duration.zero`
+
 ## 3.2.8+1
 
 * Update supported platforms in readme